refactor(etl): remove unused code from listings DWH ETL DAG
This commit removes a large block of commented-out code from the `etl_pipelines/dags/listings_dwh_etl.py` file. The removed code appears to be an older implementation of the ETL process for the listings data warehouse, which is no longer in use. Removing this dead code improves the readability and maintainability of the DAG file.